Organize git-info integration coverage (Fixes #331)
